Refactor ANSI function
